feat: extract gateway_notes_mcp and mount via capability registry
Move the notes MCP tools (get_note, create_note, delete_note) out of
gateway_oss.mcp.app into a standalone gateway_notes_mcp package,
registered as an McpCapability entry-point and mounted in
_build_mcp_app() via the capability registry before extensions run.
gateway_oss.mcp.app keeps _authenticated_clients and
register_authenticated_tools (still used for the remaining
non-notes authenticated tools and re-exported for PRO).
