+++ packages/gateway_notes_mcp/pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "gateway_notes_mcp"
+version = "0.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "gateway_notes", "fastmcp>=3.4.2"]
+
+[project.entry-points."substack_gateway.capabilities"]
+notes_mcp = "gateway_notes_mcp:capability"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_notes_mcp"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
+gateway_notes = { workspace = true }
